# backend/src/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Ventrix API",
    "version": "1.0.0",
    "description": "API de Ventrix POS y Ventrix Contable. Generada automáticamente desde src/routes (`npm run docs:openapi`), así que refleja siempre las rutas, roles y validaciones que hay en el código.\n\nLa sesión usa un access token corto (cabecera `Authorization: Bearer`) y una cookie httpOnly de refresh. Todo es multi-tenant: el negocio sale del token, nunca del cuerpo de la petición."
  },
  "servers": [
    {
      "url": "https://api.ventrix.lat/api/v1",
      "description": "Producción"
    },
    {
      "url": "http://localhost:4000/api/v1",
      "description": "Local"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Access token de un usuario (POS o Contable)."
      },
      "sellerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Sesión del portal de vendedoras (/seller), separada de la de usuarios."
      }
    },
    "schemas": {
      "RespuestaOk": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "description": "Contenido de la respuesta"
          }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "string",
            "example": "Mensaje del error"
          }
        }
      }
    }
  },
  "paths": {
...
    "/api/v1/notifications/test": {
      "post": {
        "tags": [
          "notification"
        ],
        "summary": "Envía un push de PRUEBA a los dispositivos suscritos del usuario. Sirve para que confirme, en un toque, que las notificaciones le llegan al celular.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Operación exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespuestaOk"
                }
              }
            }
          },
          "401": {
            "description": "Sin sesión o token vencido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/seller/push/test": {
      "post": {
        "tags": [
          "seller"
        ],
        "summary": "POST /seller/push/test → prueba en un toque, para que confirme que le llegan.",
        "security": [
          {
            "sellerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Operación exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespuestaOk"
                }
              }
            }
          },
          "401": {
            "description": "Sin sesión o token vencido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/users": {
      "get": {
        "tags": [
          "user"
        ],
        "summary": "CRIT-04: filtra usuarios por las sucursales del negocio autenticado",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Operación exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespuestaOk"
                }
              }
            }
          },
          "401": {
            "description": "Sin sesión o token vencido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "user"
        ],
        "summary": "CRIT-05 (POST): bloquea creación de SUPER_ADMIN por esta vía",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Operación exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespuestaOk"
                }
              }
            }
          },
          "401": {
            "description": "Sin sesión o token vencido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/{id}": {
      "delete": {
        "tags": [
          "user"
        ],
        "summary": "CRIT-05 (DELETE): verifica que el usuario target pertenezca al mismo negocio",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Operación exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespuestaOk"
                }
              }
            }
          },
          "401": {
            "description": "Sin sesión o token vencido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      },
      "patch": {
        "tags": [
          "user"
        ],
        "summary": "CRIT-05 (PATCH): verifica que el usuario target pertenezca al mismo negocio",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Operación exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespuestaOk"
                }
              }
            }
          },
          "401": {
            "description": "Sin sesión o token vencido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/api/v1/users/{id}/permissions": {
      "patch": {
        "tags": [
          "user"
        ],
        "summary": "Marcar o desmarcar permisos sueltos a un empleado. Solo se guarda lo que DIFIERE de su rol: si mañana le cambian el rol, las marcas siguen leyéndose como \"esto se lo di aparte\" y no como una copia congelada del rol viejo.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Operación exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespuestaOk"
                }
              }
            }
          },
          "401": {
            "description": "Sin sesión o token vencido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/api/v1/users/permissions/catalog": {
      "get": {
        "tags": [
          "user"
        ],
        "summary": "Catálogo de permisos + roles asignables, para dibujar la pantalla de equipo. Va antes de '/:id' y lo puede leer cualquiera que gestione usuarios.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Operación exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespuestaOk"
                }
              }
            }
          },
          "401": {
            "description": "Sin sesión o token vencido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  }
}